Move indentation tests to integration test folder.
