--- helix-core/test-indent/commands.rs
@@ -1,1 +1,0 @@
-../../helix-term/src/commands.rs
