--- helix-core/test-indent/indent.rs
@@ -1,1 +1,0 @@
-../src/indent.rs
