+++ helix-core/tests/data/indent/indent.rs
@@ -1,0 +1,1 @@
+../../../src/indent.rs
